fix: lang-js nodule_modules no longer needed
