+++ bin/lang-js/deno.json
@@ -14,6 +14,6 @@
     "fast-json-patch": "npm:fast-json-patch@3.1.1",
     "js-yaml": "npm:js-yaml@4.1.0",
     "npm:joi": "https://esm.sh/joi@17.9.1",
-    "which": "https://esm.sh/which@4.0.0"
+    "which": "https://esm.sh/which@4.0.0",
   },
 }
