+++ prelude-si/deno/BUCK
@@ -14,6 +14,10 @@ export_file(
 )
 
 export_file(
+    name = "deno_cache.py",
+)
+
+export_file(
     name = "deno_run.py",
 )
